+++ electron/main.js
@@ -128,9 +128,8 @@ function createWindow() {
   const menu = Menu.buildFromTemplate(template)
   Menu.setApplicationMenu(menu)
 
-  if (isDebug) {
+  if (isDev) {
     mainWindow.loadURL("http://localhost:3000")
-    mainWindow.webContents.openDevTools()
   } else {
     mainWindow.loadFile(path.join(__dirname, "../out/index.html"))
   }